Fix #36
If a constant was defined at top level (directly in namespace for instance), shiboken tried to generate it as a class attribute that generate build error.
This patch detect if parent MetaClass is a namespace.
If it is the case, shiboken do not generate a class attribute using writeGetterFunction but generate a module attribute using existing code defined in writeClassRegister.
See "Write static fields" block.

WARNING: global constant not defined in a namespace doesn't work yet (constant wrapper is not generated but doesn't fail)